+++ openapi/src/models/create_volume_body.rs
@@ -24,10 +24,10 @@ pub struct CreateVolumeBody {
     pub policy: crate::models::VolumeHealPolicy,
     /// number of storage replicas
     #[serde(rename = "replicas")]
-    pub replicas: i32,
+    pub replicas: u8,
     /// size of the volume in bytes
     #[serde(rename = "size")]
-    pub size: i64,
+    pub size: u64,
     /// Volume topology used to determine how to place/distribute the data.  Should either be labelled or explicit, not both.  If neither is used then the control plane will select from all available resources.
     #[serde(rename = "topology")]
     pub topology: crate::models::Topology,
@@ -37,8 +37,8 @@ impl CreateVolumeBody {
     /// CreateVolumeBody using only the required fields
     pub fn new(
         policy: impl Into<crate::models::VolumeHealPolicy>,
-        replicas: impl Into<i32>,
-        size: impl Into<i64>,
+        replicas: impl Into<u8>,
+        size: impl Into<u64>,
         topology: impl Into<crate::models::Topology>,
     ) -> CreateVolumeBody {
         CreateVolumeBody {
@@ -51,8 +51,8 @@ impl CreateVolumeBody {
     /// CreateVolumeBody using all fields
     pub fn new_all(
         policy: impl Into<crate::models::VolumeHealPolicy>,
-        replicas: impl Into<i32>,
-        size: impl Into<i64>,
+        replicas: impl Into<u8>,
+        size: impl Into<u64>,
         topology: impl Into<crate::models::Topology>,
     ) -> CreateVolumeBody {
         CreateVolumeBody {
